html/semantics/embedded-content/the-video-element/video-loading-lazy-autoplay-when-visible.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #target {
    position: absolute;
    top: 300vh;
  }
</style>

<video id="target" src="/media/A4.mp4" loading="lazy" muted playsinline autoplay width="100" height="100"></video>

<script>
  const video = document.getElementById("target");

  const t = async_test(
    "Video with loading=lazy and autoplay does not play while not visible, then plays once scrolled into view"
  );

  // Step 1: Not visible => should not play.
  t.step_timeout(() => {
    t.step(() => {
      assert_true(video.paused);
    });

    // Step 2: Scroll into view => should play.
    video.scrollIntoView();

    t.step_timeout(() => {
      t.step(() => {
        assert_false(
          video.paused,
          "Video should autoplay when scrolled into view"
        );
      });
      t.done();
    }, 1000);
  }, 1000);
</script>
</body>
</html>
